deleted unnecesary function, and refactored ifCellEmpty to be testable [
