+++ src/test/mockEvent.js
@@ -1,0 +1,17 @@
+const mockEvent = {
+    target: {
+      appendChild(imageElement) {
+          console.log(imageElement);
+        },
+        classList: {
+          add(something) {
+              console.log(something)
+          } ,
+        },
+      dataset: {
+        datasymbol: "",
+      },
+    },
+  };
+
+  export default mockEvent;
